[code-block] move code block outside of forms
