+++ src/components/forms/FormInputColor/FormInputColor_Form.tsx
@@ -5,7 +5,7 @@ import withReactContent from 'sweetalert2-react-content'
 import Swal from 'sweetalert2'
 import { FormInputColor, FormInputColorProps } from './FormInputColor'
 import * as z from 'zod'
-import { CodeBlock } from '../CodeBlock/CodeBlock'
+import { CodeBlock } from '../../CodeBlock/CodeBlock'
 import isHexColor from 'validator/lib/isHexColor'
 import { useTheme } from 'next-themes'
 import { ThemeList } from '../../../model/site/ThemeList'
