+++ src/pages/form-example/form-example-empty.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 import Swal from 'sweetalert2'
 import withReactContent from 'sweetalert2-react-content'
 import { FormExample } from '../../components/FormExample/FormExample'
-import { CodeBlock } from '../../components/forms/CodeBlock/CodeBlock'
+import { CodeBlock } from '../../components/CodeBlock/CodeBlock'
 import { Layout } from '../../components/Layout/Layout'
 import { LinksList } from '../../model/site/LinksList'
 import { checkFetchJsonResult } from '../../utils/checkFetchResult'
